third_party/blink/web_tests/wpt_internal/fenced_frame/navigate-by-name-expected.txt

This is a testharness.js-based test.
PASS setup
FAIL navigate top-level fenced frame by name assert_equals: The anchor element did not navigate a frame inside the fenced frame boundary: top-level fenced frame expected "PASS" but got "FAIL"
FAIL navigate iframe nested in a fenced frame by name assert_equals: The anchor element did not navigate a frame inside the fenced frame boundary: nested iframe expected "PASS" but got "FAIL"
FAIL navigate nested fenced frame by name assert_equals: The anchor element did not navigate a frame inside the fenced frame boundary: nested fenced frame expected "PASS" but got "FAIL"
Harness: the test ran to completion.
